deblur_sr_pipeline/deblur-sr/requirements/requirements.txt
flask==1.0.2
absl-py==0.1.9
bleach==1.5.0
click==6.7
cycler==0.10.0
decorator==4.2.1
futures==3.1.1
h5py==2.7.1
html5lib==0.9999999
imageio==2.3.0
Keras==2.2.4
Markdown==2.6.11
matplotlib==2.2.2
networkx==2.1
numpy==1.14.0
Pillow==6.2.1
protobuf==3.6.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
scikit-image==0.13.1
scipy==1.1.0
six==1.11.0
opencv-python==4.0.0.21
tqdm==4.45.0
tensorflow==1.8.0
tensorflow-tensorboard==1.5.0
Werkzeug==0.14.1
